backend/src/services/simulation_service.py
import logging
from typing import Dict, Any, Optional, List
from src.models import User
import requests
import os
from dotenv import load_dotenv

# Load environment variables
load_dotenv()

logger = logging.getLogger(__name__)

class SimulationService:
    def __init__(self):
        # Configuration for different simulation environments
        self.ros2_endpoint = os.getenv("ROS2_ENDPOINT", "http://localhost:3000")
        self.gazebo_endpoint = os.getenv("GAZEBO_ENDPOINT", "http://localhost:11345")  # Gazebo's default port
        self.unity_endpoint = os.getenv("UNITY_ENDPOINT", "http://localhost:8080")
        self.nvidia_isaac_endpoint = os.getenv("NVIDIA_ISAAC_ENDPOINT", "http://localhost:8200")

    def get_available_simulations(self) -> List[Dict[str, str]]:
        """Get list of available simulation environments"""
        simulations = [
            {
                "id": "ros2",
                "name": "ROS 2 Environment",
                "description": "Robot Operating System 2 simulation environment",
                "status": self._check_ros2_status()
            },
            {
                "id": "gazebo",
                "name": "Gazebo Simulator",
                "description": "3D dynamic simulator for robotics applications",
                "status": self._check_gazebo_status()
            },
            {
                "id": "unity",
                "name": "Unity Robotics",
                "description": "Unity 3D simulation for humanoid robotics",
                "status": self._check_unity_status()
            },
            {
                "id": "nvidia-isaac",
                "name": "NVIDIA Isaac Sim",
                "description": "NVIDIA's simulation platform for robotics",
                "status": self._check_nvidia_isaac_status()
            }
        ]
        return simulations

    def _check_ros2_status(self) -> str:
        """Check if ROS 2 environment is accessible"""
        try:
            # This is a placeholder - in a real implementation, you'd check the actual ROS 2 endpoint
            # For now, we'll return 'available' as a placeholder
            return "available"
        except Exception:
            return "unavailable"

    def _check_gazebo_status(self) -> str:
        """Check if Gazebo simulator is accessible"""
        try:
            # This is a placeholder - in a real implementation, you'd check the actual Gazebo endpoint
            return "available"
        except Exception:
            return "unavailable"

    def _check_unity_status(self) -> str:
        """Check if Unity environment is accessible"""
        try:
            # This is a placeholder - in a real implementation, you'd check the actual Unity endpoint
            return "available"
        except Exception:
            return "unavailable"

    def _check_nvidia_isaac_status(self) -> str:
        """Check if NVIDIA Isaac simulator is accessible"""
        try:
            # This is a placeholder - in a real implementation, you'd check the actual NVIDIA Isaac endpoint
            return "available"
        except Exception:
            return "unavailable"

    def launch_simulation(self, simulation_id: str, user_id: str, simulation_config: Dict[str, Any]) -> Dict[str, Any]:
        """Launch a simulation instance for a user"""
        try:
            if simulation_id == "ros2":
                return self._launch_ros2_simulation(user_id, simulation_config)
            elif simulation_id == "gazebo":
                return self._launch_gazebo_simulation(user_id, simulation_config)
            elif simulation_id == "unity":
                return self._launch_unity_simulation(user_id, simulation_config)
            elif simulation_id == "nvidia-isaac":
                return self._launch_nvidia_isaac_simulation(user_id, simulation_config)
            else:
                return {"success": False, "error": f"Unknown simulation type: {simulation_id}"}
        except Exception as e:
            logger.error(f"Error launching simulation {simulation_id}: {e}")
            return {"success": False, "error": str(e)}

    def _launch_ros2_simulation(self, user_id: str, config: Dict[str, Any]) -> Dict[str, Any]:
        """Launch a ROS 2 simulation instance"""
        # In a real implementation, this would call the ROS 2 endpoint to launch a simulation
        # For now, return a mock response
        return {
            "success": True,
            "simulation_id": f"ros2_{user_id}_{hash(str(config))}",
            "endpoint": self.ros2_endpoint,
            "status": "running",
            "connection_info": {
                "url": self.ros2_endpoint,
                "credentials": "temp_credentials"  # In real implementation, generate proper credentials
            }
        }

    def _launch_gazebo_simulation(self, user_id: str, config: Dict[str, Any]) -> Dict[str, Any]:
        """Launch a Gazebo simulation instance"""
        # In a real implementation, this would call the Gazebo endpoint to launch a simulation
        # For now, return a mock response
        return {
            "success": True,
            "simulation_id": f"gazebo_{user_id}_{hash(str(config))}",
            "endpoint": self.gazebo_endpoint,
            "status": "running",
            "connection_info": {
                "url": self.gazebo_endpoint,
                "credentials": "temp_credentials"
            }
        }

    def _launch_unity_simulation(self, user_id: str, config: Dict[str, Any]) -> Dict[str, Any]:
        """Launch a Unity simulation instance"""
        # In a real implementation, this would call the Unity endpoint to launch a simulation
        # For now, return a mock response
        return {
            "success": True,
            "simulation_id": f"unity_{user_id}_{hash(str(config))}",
            "endpoint": self.unity_endpoint,
            "status": "running",
            "connection_info": {
                "url": self.unity_endpoint,
                "credentials": "temp_credentials"
            }
        }

    def _launch_nvidia_isaac_simulation(self, user_id: str, config: Dict[str, Any]) -> Dict[str, Any]:
        """Launch a NVIDIA Isaac simulation instance"""
        # In a real implementation, this would call the NVIDIA Isaac endpoint to launch a simulation
        # For now, return a mock response
        return {
            "success": True,
            "simulation_id": f"nvidia-isaac_{user_id}_{hash(str(config))}",
            "endpoint": self.nvidia_isaac_endpoint,
            "status": "running",
            "connection_info": {
                "url": self.nvidia_isaac_endpoint,
                "credentials": "temp_credentials"
            }
        }

    def stop_simulation(self, simulation_session_id: str) -> Dict[str, Any]:
        """Stop a running simulation instance"""
        # In a real implementation, this would call the appropriate endpoint to stop the simulation
        # For now, return a mock response
        return {
            "success": True,
            "simulation_session_id": simulation_session_id,
            "status": "stopped"
        }

    def get_simulation_status(self, simulation_session_id: str) -> Dict[str, Any]:
        """Get the status of a running simulation"""
        # In a real implementation, this would check the status of the specific simulation
        # For now, return a mock response
        return {
            "simulation_session_id": simulation_session_id,
            "status": "running",
            "resources": {
                "cpu_usage": "45%",
                "memory_usage": "2.1GB",
                "gpu_usage": "60%"
            },
            "runtime": "00:15:32"
        }

    def execute_robot_command(self, simulation_session_id: str, command: str, params: Dict[str, Any]) -> Dict[str, Any]:
        """Execute a robot command in the simulation"""
        # In a real implementation, this would send the command to the appropriate simulation
        # For now, return a mock response
        return {
            "success": True,
            "simulation_session_id": simulation_session_id,
            "command": command,
            "result": "Command executed successfully",
            "execution_time": "0.123s"
        }